+++ Source_Files/Pfhortran/scripting.h
@@ -41,7 +41,7 @@ int load_script(int text_id);
 void free_script(void);
 
 // LP addition:
-int load_script_data(void *Data, int DataLen);
+int load_script_data(void *Data, size_t DataLen);
 
 // LP: don't execute initial instructions if restoring a savegame
 void script_init(bool restoring_saved);
